ベタ書きになっていたのを修正

+++ kotlin/provision/kotlin.sh
@@ -2,9 +2,14 @@
 set -Ceu
 
 USER="ocome"
+SDKMAN_HOME="/home/${USER}/.sdkman"
+SDKMAN_SHELL_PATH=${SDKMAN_HOME}"/bin/sdkman-init.sh"
+SDKMN_COMMAND="source "${SDKMAN_SHELL_PATH}
 
-su -l ${USER} -c "curl -s 'https://get.sdkman.io' | bash"
-su -l ${USER} -c "source '/home/ocome/.sdkman/bin/sdkman-init.sh'"
-su -l ${USER} -c "sdk version"
-su -l ${USER} -c "sdk install java"
-su -l ${USER} -c "sdk install kotlin"
+if [ ! -e ${SDKMAN_HOME} ]; then
+  su -l ${USER} -c "curl -s 'https://get.sdkman.io' | bash"
+  su -l ${USER} -c "${SDKMN_COMMAND}"
+  su -l ${USER} -c "sdk version"
+  su -l ${USER} -c "sdk install java"
+  su -l ${USER} -c "sdk install kotlin"
+fi
